requirements.txt
django>=4.0
pandas>=2.0
openpyxl>=3.0
djangorestframework>=3.14
